+++ tests/test_auth.py
@@ -50,7 +50,9 @@ def configuration_oauth2():
     return Configuration(
         host="host",
         api_key={
-            "Authorization": get_token({"iss": "cognito", "client_id": "cid", "exp": 0}),
+            "Authorization": get_token(
+                {"iss": "cognito", "client_id": "cid", "exp": 0}
+            ),
             "refresh": "my-refresh",
         },
     )
@@ -86,7 +88,13 @@ def test_refresh_api_key_hook_simplejwt(refresh_m, configuration_simplejwt):
 @mock.patch("threedi_api_client.auth.refresh_oauth2_token", return_value=(None, None))
 def test_refresh_api_key_hook_oauth2(refresh_m, configuration_oauth2):
     refresh_api_key(configuration_oauth2)
-    refresh_m.assert_called_once_with("cognito", configuration_oauth2)
+    refresh_m.assert_called_once_with(
+        issuer="cognito",
+        client_id="cid",
+        scope=None,
+        client_secret=None,
+        refresh_token="my-refresh",
+    )
 
 
 @mock.patch("threedi_api_client.auth.send_json_request")
@@ -134,49 +142,79 @@ def test_refresh_simplejwt_token_no_refresh(configuration_simplejwt):
 
 @mock.patch("threedi_api_client.auth.send_json_request")
 @mock.patch("threedi_api_client.auth.oauth2_autodiscovery")
-def test_refresh_oauth2_token_public(oauth2_autodiscovery, send_json_request, configuration_oauth2):
+def test_refresh_oauth2_token_public(oauth2_autodiscovery, send_json_request):
     send_json_request.return_value = {"access_token": "a"}
     oauth2_autodiscovery.return_value = {"token_endpoint": "https://authserver/token"}
 
-    access, refresh = refresh_oauth2_token("cognito", configuration_oauth2)
+    access = refresh_oauth2_token("cognito", "cid", refresh_token="my-refresh")
 
     _, kwargs = send_json_request.call_args
     assert kwargs == {
         "method": "POST",
         "url": "https://authserver/token",
-        "fields": {"grant_type": "refresh_token", "refresh_token": configuration_oauth2.api_key["refresh"], "client_id": "cid"},
+        "fields": {
+            "grant_type": "refresh_token",
+            "refresh_token": "my-refresh",
+            "client_id": "cid",
+        },
         "headers": {},
         "encode_multipart": False,
         "error_msg": "Failed to refresh the access token",
     }
 
     assert access == "a"
-    assert refresh == configuration_oauth2.api_key["refresh"]
 
 
 @mock.patch("threedi_api_client.auth.send_json_request")
 @mock.patch("threedi_api_client.auth.oauth2_autodiscovery")
-def test_refresh_oauth2_token_private(oauth2_autodiscovery, send_json_request, configuration_oauth2):
+def test_refresh_oauth2_token_private(oauth2_autodiscovery, send_json_request):
     send_json_request.return_value = {"access_token": "a"}
     oauth2_autodiscovery.return_value = {"token_endpoint": "https://authserver/token"}
-    configuration_oauth2.api_key["client_secret"] = "secret"
 
-    expected_auth_header = f"Basic {base64.b64encode(b'cid:secret').decode()}"
+    expected_auth_header = f"Basic {base64.b64encode(b'cid:my-secret').decode()}"
 
-    access, refresh = refresh_oauth2_token("cognito", configuration_oauth2)
+    access = refresh_oauth2_token(
+        "cognito", "cid", client_secret="my-secret", refresh_token="my-refresh"
+    )
 
     _, kwargs = send_json_request.call_args
     assert kwargs == {
         "method": "POST",
         "url": "https://authserver/token",
-        "fields": {"grant_type": "refresh_token", "refresh_token": configuration_oauth2.api_key["refresh"]},
+        "fields": {"grant_type": "refresh_token", "refresh_token": "my-refresh"},
         "headers": {"authorization": expected_auth_header},
         "encode_multipart": False,
         "error_msg": "Failed to refresh the access token",
     }
 
     assert access == "a"
-    assert refresh == configuration_oauth2.api_key["refresh"]
+
+
+@mock.patch("threedi_api_client.auth.send_json_request")
+@mock.patch("threedi_api_client.auth.oauth2_autodiscovery")
+def test_refresh_oauth2_token_client_credentials(
+    oauth2_autodiscovery, send_json_request
+):
+    send_json_request.return_value = {"access_token": "a"}
+    oauth2_autodiscovery.return_value = {"token_endpoint": "https://authserver/token"}
+
+    expected_auth_header = f"Basic {base64.b64encode(b'cid:my-secret').decode()}"
+
+    access = refresh_oauth2_token(
+        "cognito", "cid", client_secret="my-secret", scope="all"
+    )
+
+    _, kwargs = send_json_request.call_args
+    assert kwargs == {
+        "method": "POST",
+        "url": "https://authserver/token",
+        "fields": {"grant_type": "client_credentials", "scope": "all"},
+        "headers": {"authorization": expected_auth_header},
+        "encode_multipart": False,
+        "error_msg": "Failed to refresh the access token",
+    }
+
+    assert access == "a"
 
 
 @mock.patch("threedi_api_client.auth.send_json_request")